refactor: routers & middlewares in separate files
