+++ app.js
@@ -4,12 +4,14 @@ const bodyParser = require('body-parser');
 const helmet = require('helmet');
 const { errors } = require('celebrate');
 const authCheck = require('./middlewares/auth');
+const cors = require('./middlewares/cors');
 const authRouter = require('./routes/auth');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const { NotFoundError } = require('./utils');
 const limiter = require('./utils/rate-limiter');
 
 require('dotenv').config();
+
 const { PORT = 3000, DB_ADDRESS = 'mongodb://localhost:27017/moviesdb' } = process.env;
 
 const app = express();
@@ -21,34 +23,9 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(requestLogger);
+app.use(cors);
 
-const allowedCors = [
-  'https://bfeatb.nomoredomains.rocks',
-  'http://bfeatb.nomoredomains.rocks',
-  `http://localhost:${PORT}`,
-];
-
-app.use((req, res, next) => {
-  const { origin } = req.headers;
-  const requestHeaders = req.headers['access-control-request-headers'];
-  const DEFAULT_ALLOWED_METHODS = 'GET,HEAD,PUT,PATCH,POST,DELETE';
-
-  if (allowedCors.includes(origin)) {
-    res.header('Access-Control-Allow-Origin', origin);
-  }
-  if (req.method === 'OPTIONS') {
-    res.header('Access-Control-Allow-Methods', DEFAULT_ALLOWED_METHODS);
-    res.header('Access-Control-Allow-Headers', requestHeaders);
-    return res.end();
-  }
-  return next();
-});
-
-app.get('/crash-test', () => {
-  setTimeout(() => {
-    throw new Error('Сервер сейчас упадёт');
-  }, 0);
-});
+app.use(require('./routes/utils'));
 
 app.use(authRouter);
 app.use(authCheck);
