+++ controllers/utils.js
@@ -1,0 +1,9 @@
+function crashTest() {
+  setTimeout(() => {
+    throw new Error('Сервер сейчас упадёт');
+  }, 0);
+}
+
+module.exports = {
+  crashTest,
+};
