+++ middlewares/cors.js
@@ -1,0 +1,22 @@
+const { PORT = 3000 } = process.env;
+const allowedCors = [
+  'https://bfeatb.nomoredomains.rocks',
+  'http://bfeatb.nomoredomains.rocks',
+  `http://localhost:${PORT}`,
+];
+
+module.exports = (req, res, next) => {
+  const { origin } = req.headers;
+  const requestHeaders = req.headers['access-control-request-headers'];
+  const DEFAULT_ALLOWED_METHODS = 'GET,HEAD,PUT,PATCH,POST,DELETE';
+
+  if (allowedCors.includes(origin)) {
+    res.header('Access-Control-Allow-Origin', origin);
+  }
+  if (req.method === 'OPTIONS') {
+    res.header('Access-Control-Allow-Methods', DEFAULT_ALLOWED_METHODS);
+    res.header('Access-Control-Allow-Headers', requestHeaders);
+    return res.end();
+  }
+  return next();
+};
